Implementados algunos aspectos básicos
