+++ Camaleones/Assets/Resources/OnlinePlayer.prefab
@@ -143,6 +143,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 4541afbcf55df32a9b98ca1c0c1fc946, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  localInstance: {fileID: 0}
 --- !u!114 &7256651702470939422
 MonoBehaviour:
   m_ObjectHideFlags: 0
