+++ Camaleones/Assets/Scripts/Online/Lobby/LobbyMainMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4a36f9b505bd1b43a8fdc7b3d85ad8d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
