+++ Camaleones/Assets/Scripts/Online/Lobby/PrivateRoomCreator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ff87addd63a7a4a45824aa55efb48d07
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
