+++ Camaleones/Assets/Scripts/Online/Lobby/Room.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee714930684e97c459a7e7c1232295ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
