+++ Camaleones/Assets/Scripts/Online/ServerConstants.cs
@@ -3,4 +3,10 @@ public class ServerConstants {
     public const string GAME_VERSION = "b0.1";
     public const string ONLINE_LEVEL = "Online_Level";
 
+    #region Room Creation Constants
+
+    public const string GAME_MODE_ROOM =  "GameMode";
+
+    #endregion
+    
 }
